+++ include/zephyr/mgmt/mcumgr/transport/serial.h
