+++ include/zephyr/mgmt/mcumgr/transport/smp_bt.h
