+++ include/zephyr/mgmt/mcumgr/transport/smp_shell.h
