+++ include/zephyr/shell/shell_uart.h
@@ -10,7 +10,7 @@
 #include <zephyr/shell/shell.h>
 #include <zephyr/sys/ring_buffer.h>
 #include <zephyr/sys/atomic.h>
-#include <zephyr/mgmt/mcumgr/smp_shell.h>
+#include <zephyr/mgmt/mcumgr/transport/smp_shell.h>
 
 #ifdef __cplusplus
 extern "C" {
