+++ samples/subsys/mgmt/mcumgr/smp_svr/src/udp.c
@@ -4,10 +4,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include <zephyr/mgmt/mcumgr/smp_udp.h>
 #include <zephyr/net/net_mgmt.h>
 #include <zephyr/net/net_event.h>
 #include <zephyr/net/net_conn_mgr.h>
+#include <zephyr/mgmt/mcumgr/transport/smp_udp.h>
 
 #define LOG_LEVEL LOG_LEVEL_DBG
 #include <zephyr/logging/log.h>
