+++ subsys/mgmt/mcumgr/lib/smp/src/smp.c
@@ -12,7 +12,7 @@
 #include <string.h>
 
 #include <zephyr/net/buf.h>
-#include <zephyr/mgmt/mcumgr/smp.h>
+#include <zephyr/mgmt/mcumgr/transport/smp.h>
 #include "mgmt/mgmt.h"
 #include <zcbor_common.h>
 #include <zcbor_decode.h>
