+++ subsys/mgmt/mcumgr/smp.c
@@ -8,7 +8,7 @@
 #include <zephyr/kernel.h>
 #include <zephyr/device.h>
 #include <zephyr/net/buf.h>
-#include <zephyr/mgmt/mcumgr/smp.h>
+#include <zephyr/mgmt/mcumgr/transport/smp.h>
 #include "smp_reassembly.h"
 #include "mgmt/mgmt.h"
 #include "smp/smp.h"
