+++ subsys/mgmt/mcumgr/transport/smp_uart.c
@@ -13,10 +13,11 @@
 #include <zephyr/init.h>
 #include <zephyr/net/buf.h>
 #include <zephyr/drivers/console/uart_mcumgr.h>
+#include <zephyr/mgmt/mcumgr/transport/smp.h>
+#include <zephyr/mgmt/mcumgr/transport/serial.h>
+
 #include "mgmt/mgmt.h"
 #include "smp/smp.h"
-#include <zephyr/mgmt/mcumgr/serial.h>
-#include <zephyr/mgmt/mcumgr/smp.h>
 #include "../smp_internal.h"
 
 BUILD_ASSERT(CONFIG_MCUMGR_SMP_UART_MTU != 0, "CONFIG_MCUMGR_SMP_UART_MTU must be > 0");
