+++ tests/subsys/mgmt/mcumgr/os_mgmt_echo/src/main.c
@@ -6,8 +6,8 @@
 
 #include <zephyr/ztest.h>
 #include <zephyr/net/buf.h>
+#include <zephyr/mgmt/mcumgr/transport/smp_dummy.h>
 #include <mgmt/mgmt.h>
-#include <zephyr/mgmt/mcumgr/smp_dummy.h>
 #include <os_mgmt/os_mgmt.h>
 
 #define SMP_RESPONSE_WAIT_TIME 3
